[refactor]: 백준과 프로그래머스 이벤트 생성 함수를 별도 파일로 분리
