+++ packages/extension/src/background/index.ts
@@ -24,6 +24,7 @@ const trackers = new Map<"Programmers" | "Boj", iTracker>([
   ]))],
 ])
 
+// 백그라운드 이벤트 헨들러
 chrome.runtime.onMessage.addListener((message) => {
   const { type, payload } = message;
   if (type === undefined) {
@@ -33,11 +34,7 @@ chrome.runtime.onMessage.addListener((message) => {
   trackers.get(type)?.send(payload);
 });
 
-function handleClick() {
-  console.log("clicked");
-  // UI 열고 닫기
-}
-
+// 백준 문제 제출 감지
 chrome.webRequest.onBeforeRequest.addListener(
   ({ method, requestBody, tabId }) => {
     if (method === "GET" || requestBody?.formData === undefined) return;
@@ -66,6 +63,13 @@ chrome.webRequest.onBeforeRequest.addListener(
   },
   ["requestBody"]
 )
+
+
+// 익스텐션 아이콘 클릭시 동작
+function handleClick() {
+  console.log("clicked");
+  // UI 토글 실행
+}
 
 chrome.action.onClicked.addListener((tab) => {
   if (tab.id === undefined) return;
