+++ packages/extension/src/boj/world/index.ts
@@ -1,31 +1,8 @@
 import { eventEmitter } from "../../core/eventHub";
-import { trackerEventType } from "../../core/tracker/interface";
+import createBojTrackerEvent from "./createBojTrackerEvent";
 import pusherInterceptor from "./interceptor/pusherInterceptor";
 
 console.log("CodeVault running...");
-
-const createBojTrackerEvent = (data: string) => {
-  const { result, memory, time } = JSON.parse(data);
-
-  switch (result) {
-    case 3: // 채점 진행중
-      return {
-        type: trackerEventType.SCORE
-      }
-    case 4: // 채점 완료
-      return {
-        type: trackerEventType.SUCCESS,
-        payload: {
-          memory,
-          time
-        }
-      }
-    default: // 이외는 실패로 간주
-      return {
-        type: trackerEventType.FAIL
-      }
-  }
-}
 
 pusherInterceptor.addListener((data) => {
   eventEmitter.fromWorld({
