+++ packages/extension/src/programmers/world/index.ts
@@ -1,53 +1,24 @@
 import wsResponseBodyInterceptor from "./interceptor/wsResponseInterceptor";
 import wsRequestBodyInterceptor from "./interceptor/wsRequestInterceptor";
-import { trackerEvent, trackerEventType } from "../../core/tracker/interface";
-import { getMessageType, parseData } from "./parseGradeMessage";
+import createProgrammersTrackerEvent from "./createProgrammersTrackerEvent";
 import { eventEmitter } from "../../core/eventHub";
 
 console.log("CodeVault running...");
 
-const createEvent = (
-  messageType: Exclude<ReturnType<typeof getMessageType>, "IRRELEVANT">,
-  parsedData: ReturnType<typeof parseData>
-): trackerEvent => {
-
-  let eventType;
-  switch (messageType) {
-    case "RESULT":
-      eventType = parsedData?.passed ? trackerEventType.SUCCESS : trackerEventType.FAIL;
-      break;
-    case "SCORE":
-      eventType = trackerEventType.SCORE;
-      break;
-    case "START":
-      eventType = trackerEventType.START;
-      break;
-  }
-
-  return {
-    type: eventType,
-    payload: parsedData
-  }
-}
-
-const emitEvent = (data: string) => {
+const interceptHandler = (data: string) => {
   try {
-    const json = JSON.parse(data);
-
-    const messageType = getMessageType(json);
-    if (messageType === "IRRELEVANT") return;
-
-    const parsedData = parseData(json, messageType);
+    const programmersTrackerEvent = createProgrammersTrackerEvent(data);
+    if (programmersTrackerEvent === null) return;
 
     eventEmitter.fromWorld({
       target: "GradeTracker",
       type: "Programmers",
-      payload: createEvent(messageType, parsedData)
+      payload: programmersTrackerEvent
     });
   } catch (e) {
     console.error(e);
   }
 };
 
-wsResponseBodyInterceptor.addListener(emitEvent);
-wsRequestBodyInterceptor.addListener(emitEvent);
+wsResponseBodyInterceptor.addListener(interceptHandler);
+wsRequestBodyInterceptor.addListener(interceptHandler);
